+++ src/atv_remote/keys.py
@@ -10,16 +10,11 @@ VIRTUAL_KEYS = {
     "previous": 0xB1,
     "volume_up": 0xAF,
     "volume_down": 0xAE,
-    "up": 0x26,
-    "down": 0x28,
-    "left": 0x25,
-    "right": 0x27,
-    "select": 0x0D,  # Enter
     "back": 0x1B,  # Escape
 }
 KEYEVENTF_EXTENDEDKEY = 0x1
 KEYEVENTF_KEYUP = 0x2
-NOT_EXTENDED = {0x0D, 0x1B}
+NOT_EXTENDED = {0x1B}
 
 SUPPORTED = sys.platform == "win32"